Rename file.  `render-if.ts` -> `if.ts`
