+++ src/runtime/index.ts
@@ -3,5 +3,5 @@ export * from './index/events';
 export * from './interfaces';
 export * from './index/create-component';
 export * from './index/extract-mediator';
-export * from './index/render-if';
+export * from './index/if';
 export * from './index/each';
